SaveState: Split out Do types to reduce headers.
