+++ Core/Dialog/PSPSaveDialog.cpp
@@ -24,6 +24,7 @@
 #include "i18n/i18n.h"
 #include "thread/threadutil.h"
 
+#include "Common/FileUtil.h"
 #include "Common/ChunkFile.h"
 #include "Common/ChunkFileDo.h"
 #include "Core/FileSystems/MetaFileSystem.h"
