+++ Core/HLE/sceGe.cpp
@@ -21,6 +21,8 @@
 
 #include "Common/ChunkFile.h"
 #include "Common/ChunkFileDo.h"
+#include "Common/ChunkFileDoList.h"
+#include "Common/ChunkFileDoMap.h"
 #include "Common/ThreadSafeList.h"
 #include "Core/HLE/HLE.h"
 #include "Core/HLE/FunctionWrappers.h"
