+++ Core/HLE/sceKernelInterrupt.cpp
@@ -21,6 +21,8 @@
 
 #include "Common/ChunkFile.h"
 #include "Common/ChunkFileDo.h"
+#include "Common/ChunkFileDoList.h"
+#include "Common/ChunkFileDoMap.h"
 #include "Core/MemMapHelpers.h"
 #include "Core/Reporting.h"
 #include "Core/HLE/HLE.h"
