+++ Core/HLE/sceKernelSemaphore.cpp
@@ -18,6 +18,7 @@
 #include <algorithm>
 #include "Common/ChunkFile.h"
 #include "Common/ChunkFileDo.h"
+#include "Common/ChunkFileDoMap.h"
 #include "Core/HLE/HLE.h"
 #include "Core/MIPS/MIPS.h"
 #include "Core/CoreTiming.h"
