+++ Core/HLE/sceMp3.cpp
@@ -19,6 +19,7 @@
 #include <algorithm>
 
 #include "Common/ChunkFileDo.h"
+#include "Common/ChunkFileDoMap.h"
 #include "Core/Config.h"
 #include "Core/HLE/HLE.h"
 #include "Core/HLE/FunctionWrappers.h"
