+++ Core/MIPS/IR/IRJit.cpp
@@ -15,6 +15,7 @@
 // Official git repository and contact information can be found at
 // https://github.com/hrydgard/ppsspp and http://www.ppsspp.org/.
 
+#include <set>
 #include "base/logging.h"
 #include "ext/xxhash.h"
 #include "profiler/profiler.h"
