+++ GPU/GLES/GPU_GLES.cpp
@@ -20,6 +20,7 @@
 #include "i18n/i18n.h"
 
 #include "Common/ChunkFile.h"
+#include "Common/FileUtil.h"
 #include "Common/GraphicsContext.h"
 
 #include "Core/Config.h"
