+++ GPU/Vulkan/GPU_Vulkan.cpp
@@ -23,6 +23,7 @@
 #include "profiler/profiler.h"
 
 #include "Common/ChunkFile.h"
+#include "Common/FileUtil.h"
 #include "Common/GraphicsContext.h"
 
 #include "Core/Config.h"
